ditch other 'unused' files
